backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.28.1
python-telegram-bot==21.7
llama-cpp-python==0.3.16
